refactor: (#4) index.ts 참조하도록 파일 일괄 수정
