+++ src/pages/index.ts
@@ -1,2 +1,3 @@
 export * from './LoginPage';
+export * from './SignUpPage';
 export * from './WelcomePage';
